+++ lib/extensions/map_yaml_extension.dart
@@ -1,0 +1,7 @@
+import 'package:gits_cli/helper/helper.dart';
+
+extension MapYamlExtension on Map {
+  String get projectName => this['project_name'] ?? 'gits';
+  int get concurrent =>
+      this['concurrent'] ?? GitsModularHelper.defaultConcurrent;
+}