+++ lib/project/analyze/analyze_command.dart
@@ -1,9 +1,13 @@
 import 'package:gits_cli/constants.dart';
 import 'package:gits_cli/dependency_manager.dart';
-import 'package:gits_cli/helper/melos_helper.dart';
-import 'package:gits_cli/helper/status_helper.dart';
+import 'package:gits_cli/extensions/extensions.dart';
+import 'package:gits_cli/helper/helper.dart';
 
 class AnalyzeCommand extends Command {
+  AnalyzeCommand() {
+    argParser.addOptionGitsYaml();
+  }
+
   @override
   String get name => 'analyze';
 
@@ -14,8 +18,13 @@ class AnalyzeCommand extends Command {
   String get category => Constants.project;
 
   @override
-  void run() {
-    MelosHelper.run('melos run analyze');
+  void run() async {
+    final argGitsYaml = argResults.getOptionGitsYaml();
+
+    YamlHelper.validateGitsYaml(argGitsYaml);
+    final yaml = YamlHelper.loadFileYaml(argGitsYaml);
+
+    await GitsModularHelper.analyze(concurrent: yaml.concurrent);
     StatusHelper.success('gits_cli analyze');
   }
 }